moved jinja2 templates from dist location to v1 app path in cfgov project
