+++ cfgov/cfgov/settings/base.py
@@ -2,9 +2,7 @@ import os
 from unipath import Path
 
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-
-CFGOV_REFRESH = Path(__file__).ancestor(5).child('cfgov-refresh')
-
+SITE_ROOT = Path(__file__).ancestor(5).child('cfgov-refresh')
 GRUNT_WATCH = []
 
 SECRET_KEY = os.environ.get('SECRET_KEY', os.urandom(32))
