+++ cfgov/cfgov/settings/local.py
@@ -1,6 +1,7 @@
 from .base import *
+import v1
 
-GRUNT_WATCH = [CFGOV_REFRESH]
+GRUNT_WATCH = [SITE_ROOT]
 
 DEBUG = True
 
@@ -8,7 +9,7 @@ ALLOWED_HOSTS = ['*']
 
 STATIC_URL = '/static/'
 
-SHEER_SITES = [CFGOV_REFRESH.child('dist')]
+SHEER_SITES = [os.path.dirname(v1.__file__) + '/jinja2/v1']
 SHEER_ELASTICSEARCH_SERVER = 'localhost:9200'
 SHEER_ELASTICSEARCH_INDEX = 'content'
